refactor(runtime): centralize MF runtime plugin in zephyr-agent
