+++ libs/vite-plugin-zephyr/src/lib/internal/mf-vite-etl/runtime_plugin.mjs
@@ -1,40 +1,14 @@
 /**
- * Zephyr Runtime Plugin for Module Federation This file MUST be in ES Module
- * format (.mjs) for Vite/Rollup compatibility
+ * Zephyr Runtime Plugin for Module Federation.
  *
- * The **REMOTE_MAP** placeholder will be replaced during build time with actual
- * resolved remote dependencies by the inject_resolved_remotes function.
- *
- * IMPORTANT: Module Federation expects runtime plugins to be functions that
- * return the plugin object, not the object itself. That's why we export a
- * function here, not a plain object.
+ * Vite uses a virtual runtime plugin module serialized into remoteEntry.js by
+ * @module-federation/vite. Reuse the shared implementation from zephyr-agent.
  */
 
-const REMOTE_MAP_TEMPLATE = '"__REMOTE_MAP__"';
+import zephyrAgentModule from 'zephyr-agent';
+
+const { createZephyrRuntimePlugin } = zephyrAgentModule;
 
 export default function () {
-  return {
-    name: 'zephyr-runtime-remote-resolver',
-    beforeInit: (args) => {
-      const resolvedRemoteMap = JSON.parse(REMOTE_MAP_TEMPLATE);
-      const _windows = typeof window !== 'undefined' ? window : globalThis;
-
-      args.userOptions.remotes.forEach((remote) => {
-        const resolvedRemote = resolvedRemoteMap[remote.name];
-        if (!resolvedRemote) {
-          return;
-        }
-
-        const sessionEdgeURL = _windows.sessionStorage.getItem(
-          resolvedRemote.application_uid
-        );
-
-        const urlOverwrite = sessionEdgeURL ?? resolvedRemote.remote_entry_url;
-
-        remote.entry = urlOverwrite;
-      });
-
-      return args;
-    },
-  };
+  return createZephyrRuntimePlugin();
 }
