+++ libs/zephyr-agent/src/runtime-plugin/index.ts
@@ -1,0 +1,13 @@
+export {
+  createZephyrRuntimePlugin,
+  type ZephyrRuntimePluginOptions,
+} from './create-zephyr-runtime-plugin';
+export type {
+  BeforeRequestHookArgs,
+  FederationRuntimePlugin,
+  Options,
+  Remote,
+  RemoteWithEntry,
+} from './module-federation.types';
+
+export { createZephyrRuntimePlugin as default } from './create-zephyr-runtime-plugin';
